+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/shotlock/PrismRainCoreEntity.java
@@ -1,10 +1,6 @@
 package online.kingdomkeys.kingdomkeys.entity.shotlock;
 
 import net.minecraft.core.particles.ParticleTypes;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.syncher.EntityDataAccessor;
-import net.minecraft.network.syncher.EntityDataSerializers;
-import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
@@ -17,31 +13,19 @@ import online.kingdomkeys.kingdomkeys.entity.ModEntities;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
-public class PrismRainCoreEntity extends ThrowableProjectile {
+public class PrismRainCoreEntity extends BaseShotlockCoreEntity {
 
-	int maxTicks = 100;
-	List<RagnarokShotEntity> list = new ArrayList<RagnarokShotEntity>();
-	List<Entity> targetList = new ArrayList<Entity>();
-	float dmg;
+	List<RagnarokShotEntity> list = new ArrayList<>();
 
 	public PrismRainCoreEntity(EntityType<? extends ThrowableProjectile> type, Level world) {
 		super(type, world);
-		this.blocksBuilding = true;
+		this.maxTicks = 100;
 	}
 
 	public PrismRainCoreEntity(Level world, Player player, List<Entity> targets, float dmg) {
-		super(ModEntities.TYPE_SHOTLOCK_CIRCULAR.get(), player, world);
-		setCaster(player.getUUID());
-		String targetIDS = "";
-		for(Entity t : targets) {
-			targetIDS+=","+t.getId();
-		}
-		setTarget(targetIDS.substring(1));
-		this.targetList = targets;
-		this.dmg = dmg;
+		super(ModEntities.TYPE_SHOTLOCK_CIRCULAR.get(), world, player, targets, dmg);
+		this.maxTicks = 100;
 	}
 
 	@Override
@@ -51,7 +35,7 @@ public class PrismRainCoreEntity extends ThrowableProjectile {
 
 	@Override
 	public void tick() {
-		if (this.tickCount > maxTicks || getCaster() == null) {
+		if (isExpired()) {
 			this.remove(RemovalReason.KILLED);
 		}
 
@@ -103,23 +87,16 @@ public class PrismRainCoreEntity extends ThrowableProjectile {
 	}
 
 	private int getColor(int i) {
-		switch(i) {
-		case 0:
-			return 0xFFFFFF;
-		case 1:
-			return 0xFF0000;
-		case 2:
-			return 0x00FF00;
-		case 3:
-			return 0x0000FF;
-		case 4:
-			return 0xFF00FF;
-		case 5:
-			return 0xFFFF00;
-		case 6:
-			return 0x00FFFF;
-		}
-		return 0;
+		return switch (i) {
+			case 0 -> 0xFFFFFF;
+			case 1 -> 0xFF0000;
+			case 2 -> 0x00FF00;
+			case 3 -> 0x0000FF;
+			case 4 -> 0xFF00FF;
+			case 5 -> 0xFFFF00;
+			case 6 -> 0x00FFFF;
+			default -> 0;
+		};
 	}
 
 	@Override
@@ -127,60 +104,4 @@ public class PrismRainCoreEntity extends ThrowableProjectile {
 
 	}
 
-	public int getMaxTicks() {
-		return maxTicks;
-	}
-
-	public void setMaxTicks(int maxTicks) {
-		this.maxTicks = maxTicks;
-	}
-
-	@Override
-	public void addAdditionalSaveData(CompoundTag compound) {
-		super.addAdditionalSaveData(compound);
-		if (this.entityData.get(OWNER).isPresent()) {
-			compound.putString("OwnerUUID", this.entityData.get(OWNER).get().toString());
-			compound.putString("TargetsUUID", this.entityData.get(TARGETS));
-		}
-	}
-
-	@Override
-	public void readAdditionalSaveData(CompoundTag compound) {
-		super.readAdditionalSaveData(compound);
-		this.entityData.set(OWNER, Optional.of(UUID.fromString(compound.getString("OwnerUUID"))));
-		this.entityData.set(TARGETS, compound.getString("TargetUUID"));
-	}
-
-	private static final EntityDataAccessor<Optional<UUID>> OWNER = SynchedEntityData.defineId(PrismRainCoreEntity.class, EntityDataSerializers.OPTIONAL_UUID);
-	private static final EntityDataAccessor<String> TARGETS = SynchedEntityData.defineId(PrismRainCoreEntity.class, EntityDataSerializers.STRING);
-
-	public Player getCaster() {
-		return this.getEntityData().get(OWNER).isPresent() ? this.level().getPlayerByUUID(this.getEntityData().get(OWNER).get()) : null;
-	}
-
-	public void setCaster(UUID uuid) {
-		this.entityData.set(OWNER, Optional.of(uuid));
-	}
-
-	public List<Entity> getTargets() {
-		List<Entity> list = new ArrayList<Entity>();
-		String[] ids = this.getEntityData().get(TARGETS).split(",");
-		
-		for(String id : ids) {
-		
-			if(!id.equals(""))
-				list.add(level().getEntity(Integer.parseInt(id)));
-		}
-		return list;
-	}
-
-	public void setTarget(String lists) {
-		this.entityData.set(TARGETS, lists);
-	}
-
-	@Override
-	protected void defineSynchedData(SynchedEntityData.Builder pBuilder) {
-		pBuilder.define(OWNER, Optional.of(new UUID(0L, 0L)));
-		pBuilder.define(TARGETS, "");
-	}
 }